EWC-24 Initial Packages are installed.
