+++ packages/ext-web-components-boilerplate-create-react-app-2/.babelrc
