+++ packages/ext-web-components-boilerplate-create-react-app-2/package.json
@@ -1,0 +1,75 @@
+{
+  "name": "ext-web-components-boilerplate-create-react-app",
+  "version": "7.0.0",
+  "description": "ext-web-components-boilerplate-create-react-app",
+  "private": true,
+  "dependencies": {
+    "history": "^4.9.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-router-dom": "^5.0.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.3.4",
+    "@babel/plugin-proposal-class-properties": "^7.3.4",
+    "@babel/plugin-proposal-decorators": "^7.3.0",
+    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
+    "@babel/plugin-proposal-function-sent": "^7.2.0",
+    "@babel/plugin-proposal-json-strings": "^7.2.0",
+    "@babel/plugin-proposal-numeric-separator": "^7.2.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-syntax-import-meta": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.3.4",
+    "@babel/preset-env": "^7.3.4",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/runtime": "^7.3.4",
+    "@sencha/ext": "~7.0.0",
+    "@sencha/ext-modern": "~7.0.0",
+    "@sencha/ext-modern-theme-material": "~7.0.0",
+    "@sencha/ext-modern-treegrid": "~7.0.0",
+    "@sencha/ext-web-components": "~7.0.0",
+    "@sencha/ext-webpack-plugin": "~7.0.0",
+    "@webcomponents/webcomponentsjs": "^2.2.10",
+    "babel-eslint": "^10.0.1",
+    "base-href-webpack-plugin": "~2.0.0",
+    "copy-webpack-plugin": "^5.0.3",
+    "cross-env": "^5.2.0",
+    "css-loader": "^2.1.0",
+    "eslint": "^5.16.0",
+    "eslint-loader": "^2.1.2",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
+    "node-sass": "^4.11.0",
+    "portfinder": "^1.0.20",
+    "react-hot-loader": "^4.8.4",
+    "react-scripts": "3.0.1",
+    "rimraf": "^2.6.3",
+    "sass-loader": "^7.1.0",
+    "style-loader": "^0.23.1",
+    "webpack": "^4.29.6",
+    "webpack-cli": "^3.3.2",
+    "webpack-dev-server": "^3.2.1"
+  },
+  "scripts": {
+    "start": "webpack-dev-server --env.browser=true --env.emit=yes --env.verbose=no",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
+}
