+++ packages/ext-web-components-boilerplate-create-react-app-2/resources/favicon.ico
