+++ packages/ext-web-components-boilerplate-create-react-app-2/src/home/data.js
