+++ packages/ext-web-components-boilerplate-create-react-app-2/src/index.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import {BrowserRouter} from 'react-router-dom';
+
+import '@sencha/ext-web-components/lib/ext-button.component';
+import '@sencha/ext-web-components/lib/ext-dataview.component';
+import '@sencha/ext-web-components/lib/ext-searchfield.component';
+import '@sencha/ext-web-components/lib/ext-treelist.component';
+import '@sencha/ext-web-components/lib/ext-panel.component';
+import '@sencha/ext-web-components/lib/ext-tabpanel.component';
+import '@sencha/ext-web-components/lib/ext-titlebar.component';
+import '@sencha/ext-web-components/lib/ext-toolbar.component';
+import '@sencha/ext-web-components/lib/ext-column.component';
+import '@sencha/ext-web-components/lib/ext-grid.component';
+import '@sencha/ext-web-components/lib/ext-sheet.component';
+
+import App from './App';
+
+ReactDOM.render(<BrowserRouter><App /></BrowserRouter>, document.getElementById('root'));
