+++ packages/ext-web-components-boilerplate-create-react-app/src/index.js
@@ -1,19 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {BrowserRouter} from 'react-router-dom';
-
-import '@sencha/ext-web-components/lib/ext-button.component';
-import '@sencha/ext-web-components/lib/ext-dataview.component';
-import '@sencha/ext-web-components/lib/ext-searchfield.component';
-import '@sencha/ext-web-components/lib/ext-treelist.component';
-import '@sencha/ext-web-components/lib/ext-panel.component';
-import '@sencha/ext-web-components/lib/ext-tabpanel.component';
-import '@sencha/ext-web-components/lib/ext-titlebar.component';
-import '@sencha/ext-web-components/lib/ext-toolbar.component';
-import '@sencha/ext-web-components/lib/ext-column.component';
-import '@sencha/ext-web-components/lib/ext-grid.component';
-import '@sencha/ext-web-components/lib/ext-sheet.component';
-
+import './index.css';
 import App from './App';
+import * as serviceWorker from './serviceWorker';
 
-ReactDOM.render(<BrowserRouter><App /></BrowserRouter>, document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
+
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: https://bit.ly/CRA-PWA
+serviceWorker.unregister();
